pyproject.toml
[project]
name = "focus-gui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fitz>=0.0.1.dev2",
    "pillow>=12.0.0",
    "pymupdf>=1.26.5",
]
